@@ -4,10 +4,11 @@
   "private": true,
   "dependencies": {
     "moment": "^2.24.0",
-    "react": "^16.7.0",
+    "react": "^16.8.0-alpha.1",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^16.7.0",
+    "react-dom": "^16.8.0-alpha.1",
     "react-mentions": "^2.4.1",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3",
     "react-tippy": "^1.2.3",
     "styled-components": "^4.1.3"
